DRY root style
