+++ src/components/App.js
@@ -1,22 +1,34 @@
 import React from 'react';
 import { Provider } from 'react-redux';
+import { withStyles } from '@material-ui/core/styles';
 import { BrowserRouter as Router, Route, Switch, Redirect } from 'react-router-dom';
 import WelcomePage from '../containers/WelcomePage';
 import ChatPage from '../containers/ChatPage';
 import configureStore from '../store';
 
+const styles = theme => ({
+  root: {
+    position: 'relative',
+    display: 'flex',
+    width: '100%',
+    height: '100%',
+  },
+});
+
 const store = configureStore();
 
-const App = () => (
+const App = ({ classes }) => (
   <Provider store={store}>
     <Router>
-      <Switch>
-        <Route exact path="/(welcome)?" component={WelcomePage} />
-        <Route path="/chat" component={ChatPage} />
-        <Redirect to="/" />
-      </Switch>
+      <div className={classes.root}>
+        <Switch>
+          <Route exact path="/(welcome)?" component={WelcomePage} />
+          <Route path="/chat" component={ChatPage} />
+          <Redirect to="/" />
+        </Switch>
+      </div>
     </Router>
   </Provider>
 );
 
-export default App;
+export default withStyles(styles)(App);
