+++ src/components/ChatPage.js
@@ -1,26 +1,16 @@
 import React from 'react';
-import { withStyles } from '@material-ui/core/styles';
 import Sidebar from './Sidebar';
 import ChatHeader from './ChatHeader';
 import Chat from './Chat';
 
 import { chats, messages } from '../mock-data';
 
-const styles = theme => ({
-  root: {
-    position: 'relative',
-    display: 'flex',
-    width: '100%',
-    height: '100%',
-  },
-});
-
-const ChatPage = ({ classes }) => (
-  <div className={classes.root}>
+const ChatPage = () => (
+  <React.Fragment>
     <ChatHeader />
     <Sidebar chats={chats} />
     <Chat messages={messages} />
-  </div>
+  </React.Fragment>
 );
 
-export default withStyles(styles)(ChatPage);
+export default ChatPage;
